add cellDiameter to code generation used in dolfin-dg (using maxCellWidth for now)
Added to hack where F_c is used on the skeleton occuring if F_c involves
a gridFunction, i.e., F_c(x,u) = velocity_h*u
in this case velocity_h is not correctly restricted to ['+'/'-'] so we did
not use the correct 'x' value. I think the issue is that
F_c(x,u)['+'] is used but since 'velocity_h' in the ufl form does not involve an
explicit 'x' the ['+'] is not applied to 'velocity_h' so we end up with
an unrestricted gf there.
Currently fixed by always evaluating the coefficient on the 'inside' entity - not an
issue if the coefficient (e.g. 'velocity_h') is continuous but wrong otherwise